Fix test script for get all users endpoint
- Fix error in response body format
- Move database seeding to helper file
